Add new panes for logs, materials, market, and ship locker with associated layout updates and pane management refactoring.

+++ src/edsm/bodies.rs
@@ -26,7 +26,7 @@ pub struct Body {
     pub sub_type: Box<str>,
     pub parents: Option<Vec<HashMap<Box<str>, i32>>>,
     #[serde(rename = "distanceToArrival")]
-    pub distance_to_arrival: f64,
+    pub distance_to_arrival: f32,
     #[serde(rename = "isMainStar")]
     pub is_main_star: Option<bool>,
     #[serde(rename = "isScoopable")]
@@ -36,52 +36,52 @@ pub struct Body {
     pub spectral_class: Option<Box<str>>,
     pub luminosity: Option<Box<str>>,
     #[serde(rename = "absoluteMagnitude")]
-    pub absolute_magnitude: Option<f64>,
+    pub absolute_magnitude: Option<f32>,
     #[serde(rename = "solarMasses")]
-    pub solar_masses: Option<f64>,
+    pub solar_masses: Option<f32>,
     #[serde(rename = "solarRadius")]
-    pub solar_radius: Option<f64>,
+    pub solar_radius: Option<f32>,
     #[serde(rename = "surfaceTemperature")]
     pub surface_temperature: Option<i32>,
     #[serde(rename = "orbitalPeriod")]
-    pub orbital_period: Option<f64>,
+    pub orbital_period: Option<f32>,
     #[serde(rename = "semiMajorAxis")]
-    pub semi_major_axis: Option<f64>,
+    pub semi_major_axis: Option<f32>,
     #[serde(rename = "orbitalEccentricity")]
-    pub orbital_eccentricity: Option<f64>,
+    pub orbital_eccentricity: Option<f32>,
     #[serde(rename = "orbitalInclination")]
-    pub orbital_inclination: Option<f64>,
+    pub orbital_inclination: Option<f32>,
     #[serde(rename = "argOfPeriapsis")]
-    pub arg_of_periapsis: Option<f64>,
+    pub arg_of_periapsis: Option<f32>,
     #[serde(rename = "rotationalPeriod")]
-    pub rotational_period: Option<f64>,
+    pub rotational_period: Option<f32>,
     #[serde(rename = "rotationalPeriodTidallyLocked")]
     pub rotational_period_tidally_locked: Option<bool>,
     #[serde(rename = "axialTilt")]
-    pub axial_tilt: Option<f64>,
+    pub axial_tilt: Option<f32>,
     pub belts: Option<Vec<Belt>>,
     pub rings: Option<Vec<Ring>>,
     #[serde(rename = "updateTime")]
     pub update_time: Box<str>,
     #[serde(rename = "isLandable")]
     pub is_landable: Option<bool>,
-    pub gravity: Option<f64>,
+    pub gravity: Option<f32>,
     #[serde(rename = "earthMasses")]
-    pub earth_masses: Option<f64>,
-    pub radius: Option<f64>,
+    pub earth_masses: Option<f32>,
+    pub radius: Option<f32>,
     #[serde(rename = "surfacePressure")]
-    pub surface_pressure: Option<f64>,
+    pub surface_pressure: Option<f32>,
     #[serde(rename = "volcanismType")]
     pub volcanism_type: Option<Box<str>>,
     #[serde(rename = "atmosphereType")]
     pub atmosphere_type: Option<Box<str>>,
     #[serde(rename = "atmosphereComposition")]
-    pub atmosphere_composition: Option<HashMap<Box<str>, f64>>,
+    pub atmosphere_composition: Option<HashMap<Box<str>, f32>>,
     #[serde(rename = "solidComposition")]
-    pub solid_composition: Option<HashMap<Box<str>, f64>>,
+    pub solid_composition: Option<HashMap<Box<str>, f32>>,
     #[serde(rename = "terraformingState")]
     pub terraforming_state: Option<Box<str>>,
-    pub materials: Option<HashMap<Box<str>, f64>>,
+    pub materials: Option<HashMap<Box<str>, f32>>,
     #[serde(rename = "reserveLevel")]
     pub reserve_level: Option<Box<str>>,
 }
@@ -99,9 +99,9 @@ pub struct Belt {
     pub belt_type: Box<str>,
     pub mass: i64,
     #[serde(rename = "innerRadius")]
-    pub inner_radius: i64,
+    pub inner_radius: f32,
     #[serde(rename = "outerRadius")]
-    pub outer_radius: i64,
+    pub outer_radius: f32,
 }
 
 #[derive(Clone, Debug, Deserialize)]
@@ -111,7 +111,7 @@ pub struct Ring {
     pub ring_type: Box<str>,
     pub mass: i64,
     #[serde(rename = "innerRadius")]
-    pub inner_radius: i64,
+    pub inner_radius: f32,
     #[serde(rename = "outerRadius")]
-    pub outer_radius: i64,
+    pub outer_radius: f32,
 }
